comentarios añadidos

+++ lib/pages/cesta.dart
@@ -1,15 +1,12 @@
-// Flutter code sample for material.Card.1
-
-// This sample shows creation of a [Card] widget that shows album information
-// and two actions.
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter_login_demo/constants.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_login_demo/services/auth_provider.dart';
-import 'package:flutter_login_demo/services/authentication.dart';
 
-/// This Widget is the main application widget.
+/*
+* Clase que muestra todas las compras que ha hecho el cliente
+* y se puede borrar
+*/
 
 class Cesta extends StatefulWidget {
   @override
@@ -17,8 +14,7 @@ class Cesta extends StatefulWidget {
 }
 
 class Cestaui extends State<Cesta> {
-   
-final FirebaseAuth _auth = FirebaseAuth.instance;
+  final FirebaseAuth _auth = FirebaseAuth.instance;
   FirebaseUser user;
 
   @override
@@ -27,49 +23,48 @@ final FirebaseAuth _auth = FirebaseAuth.instance;
     initUser();
   }
 
-
   initUser() async {
     user = await _auth.currentUser();
     setState(() {});
   }
 
-  eliminarProducto(doc){
-    Firestore.instance.collection('compras').document(doc).delete();
+  // Elimina el producto y por parametro le paso el id del documento
+  eliminarProducto(doc) {
+    Firestore.instance
+        .collection(Constants.comprasCollectionId)
+        .document(doc)
+        .delete();
   }
-
-
-void mostrarDialogo(doc){
-     showDialog(
-       context: context,
-       builder: (BuildContext context) {
-         return AlertDialog(
-           title: Text('Atención'),
-           content: Text('¿Estas seguro de eliminar la compra?'),
-           actions: <Widget>[
-             FlatButton(
-               child: Text('Si'),
-               onPressed: () {
-                 Navigator.of(context).pop();
-                 eliminarProducto(doc);
-               },
-             ),
-             FlatButton(
-               child: Text('No'),
-               onPressed: () {
-                 Navigator.of(context).pop();
-               },
-             ),
-           ],
-         );
-       }
-     );
+  // muestra un dialogo que si se pulsa 'si' elimina el producto
+  void mostrarDialogo(doc) {
+    showDialog(
+        context: context,
+        builder: (BuildContext context) {
+          return AlertDialog(
+            title: Text('Atención'),
+            content: Text('¿Estas seguro de eliminar la compra?'),
+            actions: <Widget>[
+              FlatButton(
+                child: Text('Si'),
+                onPressed: () {
+                  Navigator.of(context).pop();
+                  eliminarProducto(doc);
+                },
+              ),
+              FlatButton(
+                child: Text('No'),
+                onPressed: () {
+                  Navigator.of(context).pop();
+                },
+              ),
+            ],
+          );
+        });
   }
-
-
+// muestra una lista de las compras del cliente que está conectado
   @override
   Widget build(BuildContext context) {
-
-     String email = "${user?.email}";
+    String email = "${user?.email}";
 
     return Scaffold(
       appBar: AppBar(
@@ -89,53 +84,48 @@ void mostrarDialogo(doc){
                   return new Text('Error: ${snapshot.error}');
                 switch (snapshot.connectionState) {
                   case ConnectionState.waiting:
-                        return CircularProgressIndicator();
+                    return CircularProgressIndicator();
                   case ConnectionState.none:
-                        return CircularProgressIndicator();
+                    return CircularProgressIndicator();
                   case ConnectionState.active:
-                      //  return CircularProgressIndicator();
+                  //  return CircularProgressIndicator();
                   case ConnectionState.done:
-                        return Container(
-                    child: ListView(
-              children:
-                  snapshot.data.documents.map((DocumentSnapshot document) {
-                  return Card(
-                    child: Row(
-                      children: <Widget>[
-                        Container(
-                          child: Padding(
-                            padding: EdgeInsets.all(10),
-                            child: Row(
-                              children: <Widget>[
-                                Text(
-                                  document['producto'] + '                   ',
-                                  style: TextStyle(
-                                    fontSize: 15.0,
+                    return Container(
+                        child: ListView(
+                      children: snapshot.data.documents
+                          .map((DocumentSnapshot document) {
+                        return Card(
+                          child: Row(
+                            children: <Widget>[
+                              Container(
+                                  child: Padding(
+                                padding: EdgeInsets.all(10),
+                                child: Row(children: <Widget>[
+                                  Text(
+                                    document['producto'] +
+                                        '                   ',
+                                    style: TextStyle(
+                                      fontSize: 15.0,
+                                    ),
                                   ),
-                                ),
-                                IconButton(
-                                  alignment: Alignment.centerRight,
-                                  icon: Icon(Icons.delete),
-                                  onPressed: () {
-                                     mostrarDialogo(document.documentID);
-                                  },    
-                                )
-                              ]
-                            ),
-                        )
-                        )
-                      ],
-                    ),
-                  );
-              }).toList(),
-                    )
-            );
-                  
+                                  IconButton(
+                                    alignment: Alignment.centerRight,
+                                    icon: Icon(Icons.delete),
+                                    onPressed: () {
+                                      mostrarDialogo(document.documentID);
+                                    },
+                                  )
+                                ]),
+                              ))
+                            ],
+                          ),
+                        );
+                      }).toList(),
+                    ));
                 }
               },
             )),
       ),
-
     );
   }
 }
